--- src/helpers/ft_error.c
@@ -1,23 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_error.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: yuhayrap <yuhayrap@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/16 11:36:16 by yuhayrap          #+#    #+#             */
-/*   Updated: 2024/05/16 11:36:16 by yuhayrap         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-void	ft_error(char *message, int status)
-{
-	write(2, RED, 7);
-	write(2, "Error\n", 6);
-	write(2, YELLOW, 7);
-	write(2, message, ft_strlen(message));
-	write(2, RESET_COLOR, 4);
-	exit(status);
-}
