+++ src/utils/print_arr.c
