+++ src/utils/print_commands.c
